fix: adding properties to application.properties and more exception treatment

+++ src/main/java/com/api/exercise/projectapi/service/ProjectServiceImpl.java
@@ -27,7 +27,11 @@ public class ProjectServiceImpl implements ProjectService{
 
     @Override
     public void deleteById(Long projectId) {
-        repository.deleteById(projectId);
+        if (repository.existsById(projectId)) {
+            repository.deleteById(projectId);
+        } else {
+            throw new EntityNotFoundException("Project not found");
+        }
     }
